spunoff `<Slider>`
