+++ src/App.css
@@ -139,6 +139,7 @@ fieldset .slider {
 }
 
 .slider [type='number'] {
+	background: transparent;
 	border: 0;
 	border-radius: 0;
 	height: var(--option-icon-length);
@@ -150,11 +151,14 @@ fieldset .slider {
 .menu-options label {
 	background-color: var(--color-red);
 	display: block;
+	float: left;
 	height: var(--option-icon-length);
 	padding: 0;
 	position: relative;
 	text-align: center;
 	width: var(--option-icon-length);
+	-webkit-user-select: none;
+	user-select: none;
 }
 
 .menu-options input {
